REF: BooleanArray/NumericArray share __array_ufunc__
